--- a/BotBuilder.Instrumentation.Benchmarks/Program.cs
+++ b/BotBuilder.Instrumentation.Benchmarks/Program.cs
@@ -1,0 +1,12 @@
+﻿using BenchmarkDotNet.Running;
+
+namespace BotBuilder.Instrumentation.Benchmarks
+{
+    public class Program
+    {
+        private static void Main()
+        {
+            BenchmarkRunner.Run<BotFrameworkApplicationInsightsInstrumentationTests>();
+        }
+    }
+}