--- a/botbuilder-instumentation/BotBuilder.Instrumentation.csproj
+++ b/botbuilder-instumentation/BotBuilder.Instrumentation.csproj
@@ -87,6 +87,7 @@
     <Compile Include="Dialogs\InstrumentedLuisDialog.cs" />
     <Compile Include="Instumentation\InstrumentationSettings.cs" />
     <Compile Include="Interfaces\IBotFrameworkInstrumentation.cs" />
+    <Compile Include="Interfaces\ISentimentManager.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Managers\SentimentManager.cs" />
     <Compile Include="Telemetry\BatchInput.cs" />