--- a/botbuilder-instumentation/Instumentation/InstrumentationSettings.cs
+++ b/botbuilder-instumentation/Instumentation/InstrumentationSettings.cs
@@ -1,9 +1,5 @@
 ﻿using BotBuilder.Instrumentation.Managers;
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace BotBuilder.Instrumentation.Instumentation
 {