--- a/botbuilder-instumentation/Interfaces/ISentimentManager.cs
+++ b/botbuilder-instumentation/Interfaces/ISentimentManager.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace BotBuilder.Instrumentation.Interfaces
+{
+    public interface ISentimentManager
+    {
+        Task<Dictionary<string, string>> GetSentimentProperties(string text);
+    }
+}