Added a new Console App `BotBuilder.Instrumentation.Benchmarks` We're using [BenchmarkDotNet](http://benchmarkdotnet.org/Overview.htm) library to performance test this SDK. The gist of this implementation is in `BotFrameworkApplicationInsightsInstrumentationTests.cs` For mocking BotFramework Entities we're using [Moq](https://github.com/moq/moq) framework. Benchmark results look like this;
``` ini

BenchmarkDotNet=v0.10.8, OS=Windows 10 Redstone 2 (10.0.15063)
Processor=Intel Core i7-6820HQ CPU 2.70GHz (Skylake), ProcessorCount=8
Frequency=2648437 Hz, Resolution=377.5812 ns, Timer=TSC
  [Host]     : Clr 4.0.30319.42000, 32bit LegacyJIT-v4.7.2098.0 [AttachedDebugger]
  DefaultJob : Clr 4.0.30319.42000, 32bit LegacyJIT-v4.7.2098.0

```
 |                 Method |      Mean |    Error |   StdDev |       Min |       Max |
 |----------------------- |----------:|---------:|---------:|----------:|----------:|
 | TestTrackActivityAsync | 176.02 us | 3.426 us | 5.724 us | 165.79 us | 187.63 us |
 |        TrackLuisIntent |  75.22 us | 2.207 us | 2.065 us |  73.08 us |  79.81 us |
 |          TrackQnaEvent |  75.14 us | 1.487 us | 3.356 us |  69.00 us |  84.76 us |
 |       TrackCustomEvent |  72.18 us | 1.394 us | 2.043 us |  68.55 us |  75.76 us |

Also added [AppVeyor CI pipeline](https://ci.appveyor.com/project/syedhassaanahmed/botbuilder-instrumentation-cs) which executes benchmarks on every commit and produces results as artifacts.